melvin_backend/requirements.txt
fastapi
uvicorn
websockets
python-dotenv
openai
edge-tts
pydantic
aiohttp
python-multipart
